LayoutTests/imported/w3c/web-platform-tests/user-timing/test_user_timing_mark_and_measure_exception_when_invoke_with_timing_attributes-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: wp_test

Harness Error (FAIL), message = ReferenceError: Can't find variable: wp_test

Description

This test validates exception scenarios of invoking mark() and measure() with timing attributes as value.
